pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmbox"
version = "0.1.0"
description = "Universal LLM router with free-tier-first, smart fallbacks, and Indic language support."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "LLMBox Team"}]
keywords = ["LLM", "router", "MCP", "AI", "Termux", "Indic"]
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "httpx>=0.27",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "typer>=0.12",
  "langdetect>=1.0.9",
  "cachetools>=5.3",
  "tenacity>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.23", "ruff>=0.6", "mypy>=1.10"]
cli = []
server = []

[project.scripts]
llmbox = "llmbox.cli:app"

[tool.setuptools.package-data]
llmbox = ["py.typed"]

[tool.ruff]
line-length = 100
